refactor: 경로 재설정
